aagiunta repository e classe Prenotazione (da discutere)

+++ src/main/java/com/majorbit/bozza_proj_turni_prenotazioni/domain/model/Piano.java
@@ -15,7 +15,7 @@ public class Piano {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
 
     @Column
     private String nome;
@@ -24,7 +24,7 @@ public class Piano {
     private int numero;
 
     @Column(name = "sede_id")
-    private int sedeId;
+    private long sedeId;
 
     private List<Stanza> stanze;
